- kubespray: update patches
- update dellemc.powerflex version from 1.7.0 to 2.6.0 in
pfx_requirements.yml
- ceph: upgrade version to v19.2.0
- openstack: upgrade to 2024.1 (Caracal)
- update bin.txt and images.txt
- update patch.sh script to exclude 02,03,04 patch files
- update vars.yml.sample
- Update vars files to support Rocky Linux 9
- update nova_vnc_cert submodule
- update ceph-provisioners openstack-helm chart
- ceph: bump up to v19.2.1
- ceph: bump up to v19.2.1
- update mariadb chart
- upgrade openstack-helm charts and burrito.openstack role
- upgrade kubespray to v2.28.0 and k8s v1.32.5
- add never tag to nova_vnc_cert and nova_compute_sshkey_copy playbook
since they are replaced by openstack-helm chart
- upgrade ceph_version variable to v19.2.2
- nova: add nova-compute-ssh configuration
- install helm-diff plugin for online installation
- add nova ssh public/private keys in vault
- ceph: fix templating tasks
- remove --apiserver-count from kube-apiserver manifest since it is
deprecated
- openstack(nova): add --yes-i-really-mean-it in storage init script
- openstack(openvswitch): fix --remote parameter value
- openstack(mariadb): remove backup variables
- burrito.openstack: add openvswitch in osh_infra chart group
- vars: update variables
- update binary and image list files
- update image list
- openstack(nova): fix ssh keypair variables
- vars: override coredns_deployment_nodeselector variable to run coredns
pods on control plane nodes
- prepare.sh: bump up python version to 3.12
- upgrade kubelet-csr-approver to 1.2.10
- update kubelet csr approver version in vars.yml.sample
- update download url variables
- patches: update 02, 03, 04 patch files
- patch: comment out validate in kubeadm-secondary.yml
- update genesisregistry and localrepo roles
- patch: include 02, 03, 04 patch files
- vars: update versions
- add csi-powerstore helm chart files which were missing
- update VERSIONS file
- update btx submodule to 3.0.0
- add openstack-helm git commit ID in VERSIONS
- update netapp_vars.yml
- fix(nova): add back instances_path in nova conf
- openstack(mariadb): support multiple db acl cidr
- openstack(libvirt): add kubectl image url
- openstack(mariadb): add network policy
- openstack(mariadb): add db_acl_cidrs
- openstack(mariadb): move db_acl_cidrs variable to defaults/main
- openstack(neutron): Add ovn db sync cron job
- openstack-helm: add _kubernetes_pod_priority_class.tpl in helm-toolkit
- openstack-helm: add neutron ovn db sync in configmap-bin.yaml
- openstack(neutron): add {runtime,priority}ClassName for
cron-job-ovn-db-sync
- openstack: keep the order of item list in osh_infra_charts and
osh_charts
- openstack(neutron): add neutron_ovn_db_sync image entry in
neutron.yml.j2
- openstack(ovn): Change PVC mountpoint from /etc/ovn to /var/lib/ovn
where db file is
- ansible: modify hosts.sample to update node_labels
- openstack(ovn): add ovn_controller_gw label
- openstack-helm: update version information
- mariadb: upgrade version to 10.11.13
- openstack(neutron): use public endpoint for metadata service
- openstack(nova): add security.conf in apache configuration
- openstack(neutron): add ovn_metadata_agent TLS configuration; set
ovn_db_sync mode to repair
- openstack(glance): modify useHostNetwork.api in glance.yml.j2
- openstack(neutron): metadata_proxy_shared_secret in vault
- openstack(neutron): fix tenant_network_types
- script: remove curl download in create_tarball.sh
- openstack(ovn): Set db and northd to a single replica
- descheduler: upgrade to 0.32.1
- common: add python3-libselinux in prerequiresite package
- descheduler: remove Role and RoleBinding in rbac.yaml
- descheduler: add a variable install_descheduler(default: false)
- openstack(cinder): set cinder-volume replica to 1
- openstack(mariadb): add service and mgmt cidr to mariadb netpol

+++ openstack-helm/horizon/templates/bin/_horizon.sh.tpl
@@ -18,12 +18,15 @@ set -ex
 COMMAND="${@:-start}"
 
 function start () {
-  SITE_PACKAGES_ROOT=$(python -c "from distutils.sysconfig import get_python_lib; print(get_python_lib())")
+  SITE_PACKAGES_ROOT=$(python -c "from sysconfig import get_path; print(get_path('platlib'))")
   rm -f ${SITE_PACKAGES_ROOT}/openstack_dashboard/local/local_settings.py
   ln -s /etc/openstack-dashboard/local_settings ${SITE_PACKAGES_ROOT}/openstack_dashboard/local/local_settings.py
   ln -s  ${SITE_PACKAGES_ROOT}/openstack_dashboard/conf/default_policies  /etc/openstack-dashboard/default_policies
   {{- range $key, $value := .Values.conf.horizon.local_settings_d }}
   ln -s /etc/openstack-dashboard/local_settings.d/{{ $key }}.py ${SITE_PACKAGES_ROOT}/openstack_dashboard/local/local_settings.d/{{ $key }}.py
+  {{- end }}
+  {{- range $key, $value := .Values.conf.horizon.custom_panels }}
+  ln -s /etc/openstack-dashboard/custom_panels/{{ $key }}.py ${SITE_PACKAGES_ROOT}/openstack_dashboard/local/enabled/{{ $key }}.py
   {{- end }}
   # wsgi/horizon-http needs open files here, including secret_key_store
   chown -R horizon ${SITE_PACKAGES_ROOT}/openstack_dashboard/local/
@@ -87,9 +90,15 @@ function start () {
 
   # Copy custom logo images
   {{- if .Values.manifests.configmap_logo }}
-  cp /tmp/favicon.ico ${SITE_PACKAGES_ROOT}/openstack_dashboard/static/dashboard/img/favicon.ico
-  cp /tmp/logo.svg ${SITE_PACKAGES_ROOT}/openstack_dashboard/static/dashboard/img/logo.svg
-  cp /tmp/logo-splash.svg ${SITE_PACKAGES_ROOT}/openstack_dashboard/static/dashboard/img/logo-splash.svg
+  if [[ -f /tmp/favicon.ico ]]; then
+    cp /tmp/favicon.ico ${SITE_PACKAGES_ROOT}/openstack_dashboard/static/dashboard/img/favicon.ico
+  fi
+  if [[ -f /tmp/logo.svg ]]; then
+    cp /tmp/logo.svg ${SITE_PACKAGES_ROOT}/openstack_dashboard/static/dashboard/img/logo.svg
+  fi
+  if [[ -f /tmp/logo-splash.svg ]]; then
+    cp /tmp/logo-splash.svg ${SITE_PACKAGES_ROOT}/openstack_dashboard/static/dashboard/img/logo-splash.svg
+  fi
   {{- end }}
 
   # Compress Horizon's assets.
